Change project structure (add packages), add lombok, final fields
