--- SalesAndCustomerAnalysis/src/main/java/com/innowise/Customer.java
@@ -1,59 +1,0 @@
-package com.innowise;
-
-import java.time.LocalDateTime;
-import java.util.Objects;
-
-public class Customer {
-    private String customerId;
-    private String name;
-    private String email;
-    private LocalDateTime registeredAt;
-    private int age;
-    private String city;
-
-    public Customer(String customerId, String name, String email, LocalDateTime registeredAt, int age, String city) {
-        this.customerId = customerId;
-        this.name = name;
-        this.email = email;
-        this.registeredAt = registeredAt;
-        this.age = age;
-        this.city = city;
-    }
-
-    public String getCity() {
-        return city;
-    }
-
-    public String getCustomerId() {
-        return customerId;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public LocalDateTime getRegisteredAt() {
-        return registeredAt;
-    }
-
-    public int getAge() {
-        return age;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Customer customer = (Customer) o;
-        return Objects.equals(customerId, customer.customerId);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hashCode(customerId);
-    }
-}
