--- SalesAndCustomerAnalysis/src/main/java/com/innowise/Order.java
@@ -1,40 +1,0 @@
-package com.innowise;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-public class Order {
-    private String orderId;
-    private LocalDateTime orderDate;
-    private Customer customer;
-    private List<OrderItem> items;
-    private OrderStatus status;
-
-    public Order(String orderId, LocalDateTime orderDate, Customer customer, List<OrderItem> items, OrderStatus status) {
-        this.orderId = orderId;
-        this.orderDate = orderDate;
-        this.customer = customer;
-        this.items = items;
-        this.status = status;
-    }
-
-    public Customer getCustomer() {
-        return customer;
-    }
-
-    public OrderStatus getStatus() {
-        return status;
-    }
-
-    public List<OrderItem> getItems() {
-        return items;
-    }
-
-    public String getOrderId() {
-        return orderId;
-    }
-
-    public LocalDateTime getOrderDate() {
-        return orderDate;
-    }
-}
