--- SalesAndCustomerAnalysis/src/main/java/com/innowise/OrderItem.java
@@ -1,31 +1,0 @@
-package com.innowise;
-
-public class OrderItem {
-    private String productName;
-    private int quantity;
-    private double price;
-    private Category category;
-
-    public OrderItem(String productName, int quantity, double price, Category category) {
-        this.productName = productName;
-        this.quantity = quantity;
-        this.price = price;
-        this.category = category;
-    }
-
-    public int getQuantity() {
-        return quantity;
-    }
-
-    public double getPrice() {
-        return price;
-    }
-
-    public String getProductName() {
-        return productName;
-    }
-
-    public Category getCategory() {
-        return category;
-    }
-}
