+++ SalesAndCustomerAnalysis/src/main/java/com/innowise/entity/Customer.java
@@ -1,0 +1,31 @@
+package com.innowise.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+import java.time.LocalDateTime;
+import java.util.Objects;
+
+@AllArgsConstructor
+@Getter
+public class Customer {
+    private final String customerId;
+    private final String name;
+    private final String email;
+    private final LocalDateTime registeredAt;
+    private final int age;
+    private final String city;
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Customer customer = (Customer) o;
+        return Objects.equals(customerId, customer.customerId);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hashCode(customerId);
+    }
+}
