+++ SalesAndCustomerAnalysis/src/main/java/com/innowise/enums/Category.java
@@ -1,4 +1,4 @@
-package com.innowise;
+package com.innowise.enums;
 
 public enum Category {
     ELECTRONICS, CLOTHING, BOOKS, HOME, BEAUTY, TOYS
