+++ SalesAndCustomerAnalysis/src/main/java/com/innowise/enums/OrderStatus.java
@@ -1,4 +1,4 @@
-package com.innowise;
+package com.innowise.enums;
 
 public enum OrderStatus {
     NEW, PROCESSING, SHIPPED, DELIVERED, CANCELLED
