+++ SalesAndCustomerAnalysis/src/test/java/com/innowise/analyseservice/OrderAnalysisServiceTest.java
@@ -1,10 +1,10 @@
 package com.innowise.analyseservice;
 
-import com.innowise.Category;
-import com.innowise.Customer;
-import com.innowise.Order;
-import com.innowise.OrderItem;
-import com.innowise.OrderStatus;
+import com.innowise.enums.Category;
+import com.innowise.entity.Customer;
+import com.innowise.entity.Order;
+import com.innowise.entity.OrderItem;
+import com.innowise.enums.OrderStatus;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
